+++ LibCarla/source/carla/geom/Vector2D.h
@@ -31,8 +31,6 @@ namespace geom {
 
     Vector2D() = default;
 
-    Vector2D(const Vector2D &) = default;
-
     Vector2D(float ix, float iy)
       : x(ix),
         y(iy) {}
@@ -41,18 +39,18 @@ namespace geom {
     // -- Other methods --------------------------------------------------------
     // =========================================================================
 
-    double SquaredLength() const {
+    float SquaredLength() const {
       return x * x + y * y;
     }
 
-    double Length() const {
+    float Length() const {
        return std::sqrt(SquaredLength());
     }
 
     Vector2D MakeUnitVector() const {
-      const double len = Length();
-      DEBUG_ASSERT(len > std::numeric_limits<double>::epsilon());
-      double k = 1.0 / len;
+      const float len = Length();
+      DEVELOPMENT_ASSERT(len > 2.0f * std::numeric_limits<float>::epsilon());
+      const float k = 1.0f / len;
       return Vector2D(x * k, y * k);
     }
 
@@ -82,34 +80,34 @@ namespace geom {
       return lhs;
     }
 
-    Vector2D &operator*=(const double &rhs) {
+    Vector2D &operator*=(float rhs) {
       x *= rhs;
       y *= rhs;
       return *this;
     }
 
-    friend Vector2D operator*(Vector2D lhs, const double &rhs) {
+    friend Vector2D operator*(Vector2D lhs, float rhs) {
       lhs *= rhs;
       return lhs;
     }
 
-    friend Vector2D operator*(const double &lhs, Vector2D rhs) {
+    friend Vector2D operator*(float lhs, Vector2D rhs) {
       rhs *= lhs;
       return rhs;
     }
 
-    Vector2D &operator/=(const double &rhs) {
+    Vector2D &operator/=(float rhs) {
       x /= rhs;
       y /= rhs;
       return *this;
     }
 
-    friend Vector2D operator/(Vector2D lhs, const double &rhs) {
+    friend Vector2D operator/(Vector2D lhs, float rhs) {
       lhs /= rhs;
       return lhs;
     }
 
-    friend Vector2D operator/(const double &lhs, Vector2D rhs) {
+    friend Vector2D operator/(float lhs, Vector2D rhs) {
       rhs /= lhs;
       return rhs;
     }
@@ -132,27 +130,22 @@ namespace geom {
 
 #ifdef LIBCARLA_INCLUDED_FROM_UE4
 
-    Vector2D(const FVector2D &vector)
-      : Vector2D(vector.X, vector.Y) {}
-
-    Vector2D &ToMeters(void) { // from centimeters to meters.
-       x *= 0.001f;
-       y *= 0.001f;
-       return *this;
+    /// Return a Vector2D converted from centimeters to meters.
+    Vector2D ToMeters() const {
+      return *this * 1e-2f;
     }
 
-    Vector2D &ToCentimeters(void) { // from meters to centimeters.
-       x *= 100.0f;
-       y *= 100.0f;
-       return *this;
+    /// Return a Vector2D converted from meters to centimeters.
+    Vector2D ToCentimeters() const {
+      return *this * 1e2f;
     }
 
-    operator FVector2D() const {
+    FVector2D ToFVector2D() const {
       return FVector2D{x, y};
     }
 
 #endif // LIBCARLA_INCLUDED_FROM_UE4
-  
+
     MSGPACK_DEFINE_ARRAY(x, y)
   };