+++ LibCarla/source/carla/opendrive/parser/pugixml/pugixml.cpp
@@ -40,6 +40,12 @@
 // For placement new
 #include <new>
 
+#if defined(__clang__)
+# pragma clang diagnostic push
+# pragma clang diagnostic ignored "-Wconversion"
+# pragma clang diagnostic ignored "-Wdouble-promotion"
+#endif
+
 #ifdef _MSC_VER
 #	pragma warning(push)
 #	pragma warning(disable: 4127) // conditional expression is constant
@@ -11933,7 +11939,7 @@ namespace pugi
 		assert(_result.error);
 	}
 
-	PUGI__FN const char* xpath_exception::what() const throw()
+	PUGI__FN const char* xpath_exception::what() const noexcept
 	{
 		return _result.error;
 	}
@@ -12734,6 +12740,10 @@ namespace pugi
 #endif
 
 #if defined(_MSC_VER) && defined(__c2__)
+#	pragma clang diagnostic pop
+#endif
+
+#if defined(__clang__)
 #	pragma clang diagnostic pop
 #endif