+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Recorder/CarlaReplayerHelper.cpp
@@ -4,9 +4,11 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include "CarlaReplayerHelper.h"
-#include "Carla/Actor/ActorView.h"
 #include "Carla/Actor/ActorDescription.h"
+#include "Carla/Actor/ActorView.h"
+#include "Carla/Walker/WalkerControl.h"
+#include "Carla/Walker/WalkerController.h"
+#include "CarlaReplayerHelper.h"
 
 // create or reuse an actor for replaying
 std::pair<int, FActorView>CarlaReplayerHelper::TryToCreateReplayerActor(
@@ -178,10 +180,14 @@ std::pair<int, uint32_t> CarlaReplayerHelper::ProcessReplayerEventAdd(
 
   if (result.first != 0)
   {
-    // disable physics
-    // SetActorSimulatePhysics(result.second, false);
-    // disable autopilot
-    SetActorAutopilot(result.second, false);
+    // disable physics and autopilot on vehicles
+    if (result.second.GetActorType() == FActorView::ActorType::Vehicle)
+    {
+      // disable physics
+      SetActorSimulatePhysics(result.second, false);
+      // disable autopilot
+      SetActorAutopilot(result.second, false);
+    }
   }
 
   return std::make_pair(result.first, result.second.GetActorId());
@@ -222,48 +228,33 @@ bool CarlaReplayerHelper::ProcessReplayerEventParent(uint32_t ChildId, uint32_t 
 }
 
 // reposition actors
-bool CarlaReplayerHelper::ProcessReplayerPosition(CarlaRecorderPosition Pos1, CarlaRecorderPosition Pos2, double Per)
+bool CarlaReplayerHelper::ProcessReplayerPosition(CarlaRecorderPosition Pos1, CarlaRecorderPosition Pos2, double Per, double DeltaTime)
 {
   check(Episode != nullptr);
   AActor *Actor = Episode->GetActorRegistry().Find(Pos1.DatabaseId).GetActor();
+  FVector Location;
+  FRotator Rotation;
   if (Actor  && !Actor->IsPendingKill())
   {
     // check to assign first position or interpolate between both
     if (Per == 0.0)
     {
       // assign position 1
-      FTransform Trans(FRotator::MakeFromEuler(Pos1.Rotation), FVector(Pos1.Location), FVector(1, 1, 1));
-      Actor->SetActorTransform(Trans, false, nullptr, ETeleportType::TeleportPhysics);
+      Location = FVector(Pos1.Location);
+      Rotation = FRotator::MakeFromEuler(Pos1.Rotation);
     }
     else
     {
       // interpolate positions
-      FVector Location = FMath::Lerp(FVector(Pos1.Location), FVector(Pos2.Location), Per);
-      FRotator Rotation = FMath::Lerp(FRotator::MakeFromEuler(Pos1.Rotation), FRotator::MakeFromEuler(Pos2.Rotation), Per);
-      FTransform Trans(Rotation, Location, FVector(1, 1, 1));
-      Actor->SetActorTransform(Trans, false, nullptr, ETeleportType::TeleportPhysics);
+      Location = FMath::Lerp(FVector(Pos1.Location), FVector(Pos2.Location), Per);
+      Rotation = FMath::Lerp(FRotator::MakeFromEuler(Pos1.Rotation), FRotator::MakeFromEuler(Pos2.Rotation), Per);
     }
-    // reset velocities
-    ResetVelocities(Actor);
+    // set new transform
+    FTransform Trans(Rotation, Location, FVector(1, 1, 1));
+    Actor->SetActorTransform(Trans, false, nullptr, ETeleportType::None);
     return true;
   }
   return false;
-}
-
-// reset velocity vectors on actor
-void CarlaReplayerHelper::ResetVelocities(AActor *Actor)
-{
-  if (Actor  && !Actor->IsPendingKill())
-  {
-    auto RootComponent = Cast<UPrimitiveComponent>(Actor->GetRootComponent());
-    if (RootComponent != nullptr)
-    {
-      FVector Vector(0, 0, 0);
-      // reset velocities
-      RootComponent->SetPhysicsLinearVelocity(Vector, false, "None");
-      RootComponent->SetPhysicsAngularVelocityInDegrees(Vector, false, "None");
-    }
-  }
 }
 
 // reposition the camera
@@ -308,18 +299,67 @@ bool CarlaReplayerHelper::ProcessReplayerStateTrafficLight(CarlaRecorderStateTra
   return false;
 }
 
+// set the animation for Vehicles
+void CarlaReplayerHelper::ProcessReplayerAnimVehicle(CarlaRecorderAnimVehicle Vehicle)
+{
+  check(Episode != nullptr);
+  AActor *Actor = Episode->GetActorRegistry().Find(Vehicle.DatabaseId).GetActor();
+  if (Actor && !Actor->IsPendingKill())
+  {
+    auto Veh = Cast<ACarlaWheeledVehicle>(Actor);
+    if (Veh == nullptr)
+    {
+      return;
+    }
+
+    FVehicleControl Control;
+    Control.Throttle = Vehicle.Throttle;
+    Control.Steer = Vehicle.Steering;
+    Control.Brake = Vehicle.Brake;
+    Control.bHandBrake = Vehicle.bHandbrake;
+    Control.bReverse = (Vehicle.Gear < 0);
+    Control.Gear = Vehicle.Gear;
+    Control.bManualGearShift = false;
+    Veh->ApplyVehicleControl(Control, EVehicleInputPriority::User);
+  }
+}
+
+// set the animation for walkers
+void CarlaReplayerHelper::ProcessReplayerAnimWalker(CarlaRecorderAnimWalker Walker)
+{
+  check(Episode != nullptr);
+  AActor *Actor = Episode->GetActorRegistry().Find(Walker.DatabaseId).GetActor();
+  if (Actor && !Actor->IsPendingKill())
+  {
+    auto Wal = Cast<APawn>(Actor);
+    if (Wal)
+    {
+      auto Controller = Cast<AWalkerController>(Wal->GetController());
+      if (Controller != nullptr)
+      {
+        FWalkerControl Control;
+        Control.Speed = Walker.Speed;
+        Controller->ApplyWalkerControl(Control);
+     }
+    }
+  }
+}
+
 // replay finish
 bool CarlaReplayerHelper::ProcessReplayerFinish(bool bApplyAutopilot)
 {
-  if (!bApplyAutopilot)
-  {
-    return false;
-  }
-  // set autopilot to all AI vehicles
+  // set autopilot and physics to all AI vehicles
   auto registry = Episode->GetActorRegistry();
   for (auto ActorView : registry)
   {
-    SetActorAutopilot(ActorView, true);
+    // enable physics only on vehicles
+    if (ActorView.GetActorType() == FActorView::ActorType::Vehicle)
+    {
+      SetActorSimulatePhysics(ActorView, true);
+      // autopilot
+      if (bApplyAutopilot)
+        SetActorAutopilot(ActorView, true);
+    }
   }
   return true;
 }